requirements.txt
Pillow
pyttsx3
SpeechRecognition